changes on 11 oct 2:18
