--- backend/requirements.txt
@@ -1,24 +1,0 @@
-autopep8==1.7.0
-certifi==2023.7.22
-cffi==1.15.1
-charset-normalizer==2.1.1
-click==8.1.3
-colorama==0.4.5
-Deprecated==1.2.13
-Flask==2.2.5
-gunicorn==20.1.0
-idna==3.4
-itsdangerous==2.1.2
-Jinja2==3.1.2
-MarkupSafe==2.1.1
-pycodestyle==2.9.1
-pycparser==2.21
-PyGithub==1.55
-PyJWT==2.5.0
-PyNaCl==1.5.0
-python-dotenv==0.21.0
-requests==2.31.0
-toml==0.10.2
-urllib3==1.26.12
-Werkzeug==2.2.3
-wrapt==1.14.1
